+++ feature/mission/src/main/java/com/tenday/feature/mission/components/personnel/PersonnelCardFirstHalf.kt
@@ -1,10 +1,7 @@
 package com.tenday.feature.mission.components.personnel
 
 import androidx.compose.foundation.background
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.shape.RoundedCornerShape
@@ -21,27 +18,12 @@ import com.tenday.designsystem.components.HandsUpThreeSpaceTable
 import com.tenday.designsystem.dimens.Dimens
 import com.tenday.designsystem.theme.HandsUpTypography
 import com.tenday.designsystem.theme.SingleGradientBlue
-import com.tenday.designsystem.theme.SingleGradientGray
 import com.tenday.designsystem.theme.TransparentWhite12
 import com.tenday.designsystem.theme.White
 import com.tenday.feature.mission.R
 
 @Composable
-internal fun PersonnelMissionCard(
-    modifier: Modifier = Modifier,
-) {
-    Column(
-        modifier = modifier.fillMaxWidth(),
-        verticalArrangement = Arrangement.spacedBy(Dimens.margin12),
-    ) {
-        PersonnelMissionCardFirstHalf()
-        PersonnelMissionCardSecondHalf()
-    }
-}
-
-// 상반기
-@Composable
-private fun PersonnelMissionCardFirstHalf(
+private fun PersonnelCardFirstHalf(
     modifier: Modifier = Modifier
 ) {
     HandsUpTextureCard(
@@ -90,61 +72,8 @@ private fun PersonnelMissionCardFirstHalf(
         )
     }
 }
-
-// 하반기
+@Preview(name = "PersonnelMissionCardFirstHalf")
 @Composable
-private fun PersonnelMissionCardSecondHalf(
-    modifier: Modifier = Modifier
-) {
-    HandsUpTextureCard(
-        modifier = modifier,
-        gradient = SingleGradientGray,
-        horizontalAlignment = Alignment.CenterHorizontally,
-        bottom = Dimens.margin28,
-    ) {
-        Text(
-            text = "01.01 - 06.30",
-            style = HandsUpTypography.body4.copy(
-                fontWeight = FontWeight.SemiBold,
-                color = White,
-            ),
-        )
-        Spacer(modifier = modifier.height(10.dp))
-        Text(
-            text = stringResource(R.string.mission_personnel_second_title),
-            style = HandsUpTypography.title3.copy(
-                color = White,
-            ),
-        )
-        Spacer(modifier = modifier.height(Dimens.margin6))
-        Text(
-            modifier = modifier.background(
-                color = TransparentWhite12,
-                shape = RoundedCornerShape(Dimens.cornerShape4),
-            ).padding(
-                vertical = Dimens.margin4,
-                horizontal = Dimens.margin6,
-            ),
-            text = "-",
-            style = HandsUpTypography.body3.copy(
-                fontWeight = FontWeight.SemiBold,
-                color = White,
-            ),
-        )
-        Spacer(modifier = modifier.height(Dimens.margin12))
-        HandsUpThreeSpaceTable(
-            title1 = stringResource(R.string.mission_personnel_grade_title),
-            content1 = stringResource(R.string.mission_personnel_grade_yet),
-            title2 = stringResource(R.string.mission_personnel_do_title),
-            content2 = "-",
-            title3 = stringResource(R.string.mission_personnel_analysis_title),
-            content3 = "-",
-        )
-    }
-}
-
-@Preview(name = "PersonnelMissionCard")
-@Composable
-private fun PreviewPersonnelMissionCard() {
-    PersonnelMissionCard()
+private fun PreviewPersonnelMissionCardFirstHalf() {
+    PersonnelCardFirstHalf()
 }
